feat: add test & optimize code
