+++ packages/mp-loader/package.json
@@ -4,7 +4,9 @@
   "description": "Mini-program dsl loader for Rax",
   "main": "src/index.js",
   "scripts": {
-    "start": "node ./demo/webpack-dev-server.js"
+    "start": "node ./demo/webpack-dev-server.js",
+    "test": "jest",
+    "test:watch": "npm run test -- --watchAll"
   },
   "dependencies": {
     "babel-generator": "^6.26.1",
@@ -54,6 +56,7 @@
     "webpack": "^4.28.2",
     "webpack-cli": "^3.1.2",
     "webpack-dev-server": "^3.1.10",
-    "webpack-sources": "^1.3.0"
+    "webpack-sources": "^1.3.0",
+    "jest": "^24.5.0"
   }
 }
