+++ packages/mp-loader/src/__tests__/template-loader.js
@@ -1,4 +1,4 @@
-import { runLoaders } from 'loader-runner';
+const { runLoaders } = require('loader-runner');
 
 const templateLoaderPath = require.resolve('../template-loader');
