+++ packages/mp-loader/src/adapters/__tests__/__transpileModules__/__files__/template.axml
