+++ packages/mp-loader/src/adapters/__tests__/__transpileModules__/attrs.js
@@ -1,4 +1,4 @@
-const { transformNode } = require('../attrs');
+const { transformNode } = require('../../common/transpileModules/attrs');
 
 describe('Transpile module: attrs', () => {
   it('should transform attr', () => {
