+++ packages/mp-loader/src/adapters/alipay/transpileModules/list.js
@@ -1,4 +1,4 @@
-const { transformNodeWrapper} = require('../../../transpiler/transpileModules/list');
+const { createTransformNode } = require('../../../transpiler/transformCreator/list');
 
 const ATTRIBUTES = {
   FOR: 'a:for',
@@ -7,5 +7,5 @@ const ATTRIBUTES = {
 };
 
 module.exports = {
-  transformNode: transformNodeWrapper(ATTRIBUTES)
+  transformNode: createTransformNode(ATTRIBUTES)
 };
