+++ packages/mp-loader/src/adapters/common/transpileModules/style.js
@@ -1,0 +1,7 @@
+const { createTransformNode, staticKeys, genData } = require('../../../transpiler/transformCreator/style');
+
+module.exports = {
+  transformNode: createTransformNode(),
+  staticKeys,
+  genData
+};
