+++ packages/mp-loader/src/adapters/common/transpileModules/tagName.js
@@ -1,0 +1,5 @@
+const { createTransformNode } = require('../../../transpiler/transformCreator/tagName');
+
+module.exports = {
+  transformNode: createTransformNode()
+};
