+++ packages/mp-loader/src/adapters/weapp/transpileModules/bind.js
@@ -1,7 +1,7 @@
-const { transformNodeWrapper} = require('../../../transpiler/transpileModules/bind');
+const { createTransformNode } = require('../../../transpiler/transformCreator/bind');
 
 const IS_DETECTIVE = /^wx\:/;
 
 module.exports = {
-  transformNode: transformNodeWrapper(IS_DETECTIVE)
+  transformNode: createTransformNode(IS_DETECTIVE)
 };
