+++ packages/mp-loader/src/adapters/weapp/transpileModules/events.js
@@ -1,7 +1,7 @@
-const { transformNodeWrapper} = require('../../../transpiler/transpileModules/events');
+const { createTransformNode } = require('../../../transpiler/transformCreator/events');
 
 const LISTENER_ACTION = 'bind';
 
 module.exports = {
-  transformNode: transformNodeWrapper(LISTENER_ACTION)
+  transformNode: createTransformNode(LISTENER_ACTION)
 };
