+++ packages/mp-loader/src/transpiler/createParse.js
@@ -1,6 +1,6 @@
 const { parse, baseOptions } = require('sfc-compiler');
 
-exports.createParse = function(modules) {
+module.exports = function(modules) {
   return function(ast, opt) {
     const options = Object.assign({}, baseOptions, { modules }, opt);
     return parse(ast, options);
